aws-mwaa-local-runner/requirements/requirements.txt
pandas==1.5.1
requests==2.28
python-dotenv==0.21
psycopg2==2.9
glom==22.1
kafka-python==2.0
cmake
apache-airflow
airflow-provider-kafka
pendulum